+++ src/main/kotlin/lotto/model/LottoMachine.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 class LottoMachine {
