+++ src/main/kotlin/lotto/model/LottoNumbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 data class LottoNumbers(
     private val lottoNumbers: List<LottoNumber>
