+++ src/main/kotlin/lotto/model/LottoStore.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 class LottoStore {
