+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -1,0 +1,3 @@
+package lotto.view
+
+class OutputView
