+++ src/test/kotlin/lotto/Fixtures.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.model.LottoNumber
+
 object Fixtures {
     fun createSixLottoNumber(numbers: List<Int>): List<LottoNumber> {
         return numbers.map { LottoNumber.from(it) }
