+++ src/test/kotlin/lotto/LottoNumbersTest.kt
@@ -1,6 +1,7 @@
 package lotto
 
 import lotto.Fixtures.createSixLottoNumber
+import lotto.model.LottoNumbers
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
