+++ src/test/kotlin/lotto/LottoStoreTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.model.LottoStore
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
